Change data type of Version from Long to Int

+++ KvStore.Api/Controllers/KeyValueController.cs
@@ -39,7 +39,7 @@ public sealed class KeyValueController(
     public async Task<ActionResult<KeyValueResponse>> PutAsync(
         string key,
         [FromBody] JsonNode? value,
-        [FromQuery(Name = "ifVersion")] long? expectedVersion,
+        [FromQuery(Name = "ifVersion")] int? expectedVersion,
         CancellationToken cancellationToken)
     {
         var response = await commandDispatcher.DispatchAsync(
@@ -55,7 +55,7 @@ public sealed class KeyValueController(
     public async Task<ActionResult<KeyValueResponse>> PatchAsync(
         string key,
         [FromBody] JsonNode? delta,
-        [FromQuery(Name = "ifVersion")] long? expectedVersion,
+        [FromQuery(Name = "ifVersion")] int? expectedVersion,
         CancellationToken cancellationToken)
     {
         var response = await commandDispatcher.DispatchAsync(
