+++ KvStore.Core/Application/KeyValue/Commands/PutKeyValue/PutKeyValueCommand.cs
@@ -4,6 +4,6 @@ using KvStore.Core.Application.KeyValue.Responses;
 
 namespace KvStore.Core.Application.KeyValue.Commands.PutKeyValue;
 
-public sealed record PutKeyValueCommand(string Key, JsonNode? Value, long? ExpectedVersion)
+public sealed record PutKeyValueCommand(string Key, JsonNode? Value, int? ExpectedVersion)
     : ICommand<KeyValueResponse>;
