+++ KvStore.Router/Models/KeyValueRecord.cs
@@ -2,5 +2,5 @@ using System.Text.Json.Nodes;
 
 namespace KvStore.Router.Models;
 
-public sealed record KeyValueRecord(string Key, JsonNode? Value, long Version);
+public sealed record KeyValueRecord(string Key, JsonNode? Value, int Version);
